Add categoryService error return. Add pointer (created_at, updated_at). Update the returned http code. Adjust int to int64.

+++ request/publicRequest.go
@@ -8,5 +8,5 @@ type Pagination struct {
 }
 
 type TableID struct {
-	Id int `uri:"id" binding:"required"`
+	Id int64 `uri:"id" binding:"required"`
 }
